refactored ir code structure
